Moved regexps data into sls-utils
